Split native Zenmap output text view helpers
